requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
websockets>=12.0
pydantic-settings>=2.3.0
pytest>=8.2.0
httpx>=0.27.0
